requirements.txt
# データ処理
pandas>=2.1.0
openpyxl>=3.1.0
numpy>=1.26.0

# Google Sheets連携
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Neo4j連携
neo4j>=5.14.0

# 設定・環境変数
pyyaml>=6.0
python-dotenv>=1.0.0

# HTTP通信
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# ユーティリティ
uuid>=1.30
python-dateutil>=2.8.0

# ログ・デバッグ
loguru>=0.7.0

# テスト
pytest>=7.4.0
pytest-cov>=4.1.0

# データ検証
pydantic>=2.5.0

# CLI
click>=8.1.0
rich>=13.7.0

# バックエンド (FastAPI)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# フロントエンド (Streamlit)
streamlit>=1.31.0

# ドキュメント生成
reportlab>=4.0.0
python-docx>=1.1.0
Pillow>=10.0.0

# LLM統合 (Ollama + LangChain)
langchain>=0.1.0
langchain-community>=0.0.17
httpx>=0.26.0

# 開発ツール
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0
isort>=5.13.0
